Moved mac library to new directory structure and got shared tests running.

+++ shared/test/NSDate_ConventionalC_SharedTests.m
@@ -10,17 +10,17 @@
 
 -(void)testDateSecond
 {
-    STAssertEquals(self.date.second, 58, nil);
+    STAssertEquals((int)self.date.second, 58, nil);
 }
 
 -(void)testDateMinute
 {
-    STAssertEquals(self.date.minute, 3, nil);
+    STAssertEquals((int)self.date.minute, 3, nil);
 }
 
 -(void)testDateHour
 {
-    STAssertEquals(self.date.hour, 19, nil);
+    STAssertEquals((int)self.date.hour, 19, nil);
 }
 
 @end
